+++ lib/public_suffix/list.rb
@@ -195,6 +195,9 @@ module PublicSuffix
     #
     # @return [Array<PublicSuffix::Rule::*>]
     def select(domain)
+      # raise DomainInvalid, "`#{domain}' is not expected to contain a scheme"
+      return [] if domain.include?("://")
+
       indices = (@indexes[Domain.domain_to_labels(domain).first] || [])
       @rules.values_at(*indices).select { |rule| rule.match?(domain) }
     end