+++ test/acceptance_test.rb
@@ -20,16 +20,19 @@ class AcceptanceTest < Test::Unit::TestCase
       assert_equal tld, domain.tld, "Invalid tld for '#{name}'"
       assert_equal sld, domain.sld, "Invalid sld for '#{name}'"
       assert_equal trd, domain.trd, "Invalid trd for '#{name}'"
+      assert PublicSuffix.valid?(name)
     end
   end
 
   InvalidCases = {
     "nic.ke" => PublicSuffix::DomainNotAllowed,
+    "http://www.google.com" => PublicSuffix::DomainInvalid,
   }
 
   def test_invalid
     InvalidCases.each do |name, error|
       assert_raise(error) { PublicSuffix.parse(name) }
+      assert !PublicSuffix.valid?(name)
     end
   end